+++ consensus/block-storage-proto/src/proto/block_storage.proto
@@ -4,44 +4,17 @@ package block_storage;
 
 import "consensus.proto";
 import "google/protobuf/empty.proto";
+import "explorer.proto";
 
 // --------------------------------------------------------------
 // ----------- Service definition for BlockStorage
 // --------------------------------------------------------------
 service BlockStorage {
-    rpc GetBlockByBlockId(BlockId) returns (GetBlockByBlockIdResponse);
-    rpc GetBlockSummaryList(GetBlockSummaryListRequest) returns (GetBlockSummaryListResponse);
-    rpc LatestBlockHeight(google.protobuf.Empty) returns (LatestBlockHeightResponse);
+    rpc GetBlockByBlockId(types.BlockId) returns (GetBlockByBlockIdResponse);
+    rpc GetBlockSummaryList(types.GetBlockSummaryListRequest) returns (types.GetBlockSummaryListResponse);
+    rpc LatestBlockHeight(google.protobuf.Empty) returns (types.LatestBlockHeightResponse);
 }
 
 message GetBlockByBlockIdResponse {
     consensus.Block block = 1;
-}
-
-message BlockSummary {
-    bytes block_id = 1;
-    uint64 height = 2;
-    bytes parent_id = 3;
-    bytes accumulator_root_hash = 4;
-    bytes state_root_hash = 5;
-    bytes miner = 6;
-    uint64 nonce = 7;
-    bytes target = 8;
-    uint32 algo = 9;
-}
-
-message GetBlockSummaryListRequest {
-    BlockId block_id = 1;
-}
-
-message BlockId {
-    bytes id = 1;
-}
-
-message GetBlockSummaryListResponse {
-    repeated BlockSummary blocks = 1;
-}
-
-message LatestBlockHeightResponse {
-    uint64 height = 1;
 }
